M3U Support
Images are now embedded in the script. Added an option to put multi-disc ROMs in a .m3u folder so Emulation Station views it as a single game. As far as I know, PS1, Dreamcast, and Saturn emulators are the only ones that support .m3us, so it is restricted to them.